@@ -43,27 +43,27 @@
     "@swc/jest": "^0.2.15",
     "@testing-library/dom": "^8.11.1",
     "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/react-hooks": "^7.0.2",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/react": "^13.1.1",
+    "@testing-library/react-hooks": "^8.0.0",
+    "@testing-library/user-event": "^14.1.1",
     "@types/jest": "^27.0.3",
     "@types/lodash": "^4.14.178",
-    "@types/node": "^16.11.12",
+    "@types/node": "^17.0.25",
     "@types/papaparse": "^5.3.2",
-    "@types/react": "^17.0.37",
+    "@types/react": "^18.0.5",
     "@types/recharts": "^1.8.23",
     "@typescript-eslint/eslint-plugin": "^5.7.0",
     "@typescript-eslint/parser": "^5.7.0",
     "eslint": "^8.4.1",
     "eslint-config-google": "^0.14.0",
-    "eslint-plugin-jest-dom": "^3.9.2",
+    "eslint-plugin-jest-dom": "^4.0.1",
     "eslint-plugin-react": "^7.27.1",
     "i18n-unused": "^0.8.0",
     "jest": "^27.4.5",
-    "msw": "^0.36.3",
+    "msw": "^0.39.2",
     "msw-storybook-addon": "^1.5.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "ts-jest": "^27.1.1",
     "tsconfig-paths-webpack-plugin": "^3.5.2"
   },
@@ -77,8 +77,8 @@
   },
   "dependencies": {
     "@bugsnag/js": "^7.16.2",
-    "@types/react-dom": "^17.0.9",
-    "graphql": "16.0.0",
+    "@types/react-dom": "^18.0.1",
+    "graphql": "^16.3.0",
     "lerna": "^4.0.0",
     "papaparse": "^5.3.2"
   },
