@@ -20,21 +20,21 @@
     "date-fns": "^2.27.0",
     "dompurify": "^2.3.5",
     "graphql-import-node": "^0.0.4",
-    "graphql-request": "^3.7.0",
+    "graphql-request": "^4.2.0",
     "history": "^5.1.0",
-    "i18next": "^21.6.0",
-    "i18next-browser-languagedetector": "^6.1.2",
+    "i18next": "^21.6.16",
+    "i18next-browser-languagedetector": "^6.1.4",
     "i18next-chained-backend": "^3.0.2",
-    "i18next-http-backend": "^1.3.1",
+    "i18next-http-backend": "^1.4.0",
     "i18next-localstorage-backend": "^3.1.3",
     "js-cookie": "^3.0.1",
-    "kbar": "^0.1.0-beta.27",
+    "kbar": "^0.1.0-beta.33",
     "lodash": "^4.17.21",
     "notistack": "^2.0.3",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-currency-input-field": "^3.6.0",
-    "react-dom": "^17.0.2",
-    "react-i18next": "^11.15.1",
+    "react-dom": "^18.0.0",
+    "react-i18next": "^11.16.7",
     "react-query": "^3.34.12",
     "react-router-dom": "^6.1.1",
     "recharts": "^2.1.9",
@@ -52,7 +52,7 @@
     "css-mediaquery": "^0.1.2",
     "eslint-config-prettier": "^8.3.0",
     "husky": "^7.0.4",
-    "lint-staged": "^11.1.1",
+    "lint-staged": "^12.3.8",
     "prettier": "^2.5.1",
     "typescript": "^4.5.4"
   }
