@@ -6,7 +6,7 @@
   "devDependencies": {
     "eslint-config-prettier": "^8.3.0",
     "husky": "^7.0.1",
-    "lint-staged": "^11.1.1",
+    "lint-staged": "^12.3.8",
     "prettier": "^2.3.2",
     "typescript": "^4.3.5"
   },
@@ -18,8 +18,8 @@
     "@openmsupply-client/config": "^0.0.0",
     "@openmsupply-client/invoices": "^0.0.0",
     "@openmsupply-client/requisitions": "^0.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "recharts": "1.8.5"
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "recharts": "^2.1.9"
   }
 }
