@@ -4,20 +4,20 @@
   "main": "./src/index.ts",
   "private": true,
   "devDependencies": {
-    "copy-webpack-plugin": "^9.0.1",
+    "copy-webpack-plugin": "^10.2.4",
     "eslint-config-prettier": "^8.3.0",
-    "find-up": "5.0.0",
-    "html-webpack-plugin": "5.3.1",
+    "find-up": "^6.3.0",
+    "html-webpack-plugin": "^5.5.0",
     "husky": "^7.0.1",
-    "lint-staged": "^11.1.1",
+    "lint-staged": "^12.3.8",
     "prettier": "^2.3.2",
-    "react-refresh": "^0.10.0",
+    "react-refresh": "^0.12.0",
     "rimraf": "^3.0.2",
-    "serve": "11.3.2",
+    "serve": "^13.0.2",
     "typescript": "^4.3.5",
-    "webpack": "5.40.0",
-    "webpack-cli": "4.9.0",
-    "webpack-dev-server": "4.0.0"
+    "webpack": "^5.72.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.8.1"
   },
   "scripts": {
     "start": "webpack-cli serve",
@@ -32,15 +32,15 @@
     "@openmsupply-client/config": "^0.0.0",
     "@openmsupply-client/dashboard": "^0.0.0",
     "@openmsupply-client/inventory": "^0.0.0",
+    "@openmsupply-client/invoices": "^0.0.0",
     "@openmsupply-client/requisitions": "^0.0.0",
     "@openmsupply-client/system": "^0.0.0",
-    "@openmsupply-client/invoices": "^0.0.0",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.4",
     "history": "^5.1.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-router-dom": "^6.0.2",
-    "swc-loader": "^0.1.14",
+    "swc-loader": "^0.2.0",
     "webpack-bundle-analyzer": "^4.4.2"
   },
   "msw": {
