@@ -6,7 +6,7 @@
   "devDependencies": {
     "eslint-config-prettier": "^8.3.0",
     "husky": "^7.0.1",
-    "lint-staged": "^11.1.1",
+    "lint-staged": "^12.3.8",
     "prettier": "^2.3.2",
     "typescript": "^4.3.5"
   },
@@ -17,7 +17,7 @@
     "@openmsupply-client/common": "^0.0.1",
     "@openmsupply-client/config": "^0.0.0",
     "@openmsupply-client/requisitions": "^0.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
   }
 }
