@@ -1,4 +1,4 @@
-import React, { FC } from 'react';
+import React, { FC, PropsWithChildren } from 'react';
 import {
   MenuItem,
   MenuItemProps,
@@ -75,7 +75,7 @@ interface DropdownMenuProps {
 
 // Styled doesn't like `sx` prop being passed to it.
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-export const DropdownMenu: FC<DropdownMenuProps> = ({
+export const DropdownMenu: FC<PropsWithChildren<DropdownMenuProps>> = ({
   label,
   children,
   disabled = false,
