@@ -5,8 +5,11 @@ import {
   ConfirmationModalControllerState,
 } from './ConfirmationModalContext';
 import { ConfirmationModal } from './ConfirmationModal';
+import { PropsWithChildrenOnly } from '@common/types';
 
-export const ConfirmationModalProvider: FC = ({ children }) => {
+export const ConfirmationModalProvider: FC<PropsWithChildrenOnly> = ({
+  children,
+}) => {
   const [confirmationModalState, setState] = useState<ConfirmationModalState>({
     open: false,
     message: '',
