@@ -1,4 +1,4 @@
-import React, { FC } from 'react';
+import React, { FC, PropsWithChildren } from 'react';
 import Fade from '@mui/material/Fade';
 import Popper, { PopperProps } from '@mui/material/Popper';
 import { useAppTheme } from '@common/styles';
@@ -10,7 +10,7 @@ export interface BasePopoverProps extends Omit<PopperProps, 'open'> {
   height?: number;
 }
 
-export const BasePopover: FC<BasePopoverProps> = ({
+export const BasePopover: FC<PropsWithChildren<BasePopoverProps>> = ({
   children,
   isOpen,
   anchorEl,
