@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { PropsWithChildren } from 'react';
 import { usePopover } from '../BasePopover';
 
 interface PopoverProps {
@@ -6,7 +6,7 @@ interface PopoverProps {
   placement?: 'top' | 'bottom' | 'left' | 'right';
 }
 
-export const HoverPopover: React.FC<PopoverProps> = ({
+export const HoverPopover: React.FC<PropsWithChildren<PopoverProps>> = ({
   children,
   Content,
   placement = 'left',
