@@ -1,4 +1,4 @@
-import React, { FC } from 'react';
+import React, { FC, PropsWithChildren } from 'react';
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 
@@ -8,12 +8,9 @@ export interface PaperPopoverSectionProps {
   alignItems?: 'center' | 'flex-start' | 'flex-end' | 'stretch';
 }
 
-export const PaperPopoverSection: FC<PaperPopoverSectionProps> = ({
-  children,
-  label,
-  labelStyle,
-  alignItems,
-}) => (
+export const PaperPopoverSection: FC<
+  PropsWithChildren<PaperPopoverSectionProps>
+> = ({ children, label, labelStyle, alignItems }) => (
   <Box
     gap={2}
     p={3}
