@@ -4,6 +4,7 @@ import { renderHook } from '@testing-library/react-hooks';
 import { useIsGrouped } from './hooks';
 import { TableStore, TableProvider, createTableStore } from './TableContext';
 import { LocalStorage } from '../../../../localStorage';
+import { PropsWithChildrenOnly } from '@common/types';
 
 const useStore = createTableStore();
 
@@ -298,7 +299,7 @@ describe('TableContext - grouping rows', () => {
 });
 
 describe('useIsGrouped', () => {
-  const Wrapper: FC = ({ children }) => {
+  const Wrapper: FC<PropsWithChildrenOnly> = ({ children }) => {
     return (
       <TableProvider createStore={createTableStore}>{children}</TableProvider>
     );
