@@ -3,8 +3,6 @@ import Host from './Host';
 import '@fontsource/inter/variable.css';
 // import "@fontsource/inter/variable-italic.css"; // Italic variant: not currently used
 
-const App: FC = () => {
-  return <Host />;
-};
+const App: FC = () => <Host />;
 
 export default App;
