@@ -4,5 +4,5 @@ import { StocktakeLineFragment } from './Stocktake/api/operations.generated';
 export interface StocktakeSummaryItem {
   id: string;
   item?: ItemRowFragment | null;
-  lines: [StocktakeLineFragment, ...StocktakeLineFragment[]];
+  lines: StocktakeLineFragment[];
 }
